🚚 Put server admin queries in one file
