+++ frontend/app/lib/queries/index.ts
@@ -6,11 +6,10 @@ export * from "./logs";
 export * from "./misc";
 export * from "./preview-templates";
 export * from "./projects";
-export * from "./registries";
 export * from "./server";
 export * from "./services";
 export * from "./shared";
-export * from "./ssh-keys";
+export * from "./server-admin";
 export * from "./templates";
 export * from "./users";
 export * from "./workspaces";
