+++ frontend/app/lib/queries/server-admin.ts
@@ -163,3 +163,24 @@ export const buildRegistryQueries = {
       placeholderData: keepPreviousData
     })
 };
+
+export const sshKeysQueries = {
+  list: queryOptions({
+    queryKey: ["SSH_KEYS"] as const,
+    queryFn: async ({ signal }) => {
+      const { data } = await apiClient.GET("/api/shell/ssh-keys/", {
+        signal
+      });
+      if (!data) {
+        throw notFound("Oops !");
+      }
+      return data;
+    },
+    refetchInterval: (query) => {
+      if (query.state.data) {
+        return DEFAULT_QUERY_REFETCH_INTERVAL;
+      }
+      return false;
+    }
+  })
+};
